% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{go_create_description}
\alias{go_create_description}
\title{Create description tibble}
\usage{
go_create_description()
}
\value{
Tibble containing descriptions of GO IAQS scores
}
\description{
Create description tibble
}
\examples{
description_tb <- go_create_description()
}
